Add userRoute to server

+++ server.js
@@ -1,11 +1,14 @@
 import path from 'path'
 import express from 'express'
 import cookieParser from 'cookie-parser'
-import { bugService} from './services/bug.service.js'
+
+import { bugService } from './services/bug.service.js'
 import { loggerService } from './services/logger.service.js'
+import { userService } from './services/user.service.js'
+import { authService } from './services/auth.servic.js'
 
 const app = express()
- 
+
 //* Express Config:
 app.use(express.static('public'))
 app.use(express.json())
@@ -18,11 +21,11 @@ app.get('/api/bug', (req, res) => {
     const filterBy = {
         txt: req.query.txt || '',
         minSeverity: +req.query.minSeverity || 0,
-        pageIdx: req.query.pageIdx ,
+        pageIdx: req.query.pageIdx,
         sortBy: req.query.sortBy || '',
         sortDir: +req.query.sortDir || 1,
         labels: req.query.labels || []
-}
+    }
     bugService.query(filterBy)
         .then(bugs => res.send(bugs))
         .catch(err => {
@@ -33,8 +36,10 @@ app.get('/api/bug', (req, res) => {
 
 //* Create
 app.post('/api/bug', (req, res) => {
-    const bugToSave = req.body
+    const loggedinUser = authService.validateToken(req.cookies.loginToken)
+    if (!loggedinUser) return res.status(401).send(`Can't add bug`)
 
+    const bugToSave = req.body
     bugService.save(bugToSave)
         .then(bug => res.send(bug))
         .catch(err => {
@@ -46,8 +51,10 @@ app.post('/api/bug', (req, res) => {
 
 //*Edit
 app.put('/api/bug/:bugId', (req, res) => {
-    const bugToSave = req.body
+    const loggedinUser = authService.validateToken(req.cookies.loginToken)
+    if (!loggedinUser) return res.status(401).send(`Can't update bug`)
 
+    const bugToSave = req.body
     bugService.save(bugToSave)
         .then(bug => res.send(bug))
         .catch(err => {
@@ -61,8 +68,8 @@ app.get('/api/bug/:bugId', (req, res) => {
     const { bugId } = req.params
     const { visitedBugs = [] } = req.cookies
 
-    if(visitedBugs.length >= 3) return res.status(401).send('Wait for a bit')
-    if(!visitedBugs.includes(bugId)) visitedBugs.push(bugId)
+    if (visitedBugs.length >= 3) return res.status(401).send('Wait for a bit')
+    if (!visitedBugs.includes(bugId)) visitedBugs.push(bugId)
 
     res.cookie('visitedBugs', visitedBugs, { maxAge: 1000 * 70 })
 
@@ -76,6 +83,9 @@ app.get('/api/bug/:bugId', (req, res) => {
 
 //* Remove/Delete
 app.delete('/api/bug/:bugId', (req, res) => {
+    const loggedinUser = authService.validateToken(req.cookies.loginToken)
+    if (!loggedinUser) return res.status(401).send(`Can't remove bug`)
+
     const { bugId } = req.params
     bugService.remove(bugId)
         .then(() => res.send('Bug Removed'))
@@ -85,9 +95,64 @@ app.delete('/api/bug/:bugId', (req, res) => {
         })
 })
 
- app.get('/', (req, res) => res.send('Hello there')) 
- app.get('/**', (req, res) => {
-     res.sendFile(path.resolve('public/index.html'))
- })
- 
- app.listen(3030, () => console.log('Server ready at port 3030'))
+// User API
+app.get('/api/user', (req, res) => {
+    userService.query()
+        .then(users => res.send(users))
+        .catch(err => {
+            loggerService.error('Cannot load users', err)
+            res.status(400).send('Cannot load users')
+        })
+})
+
+app.get('/api/user/:userId', (req, res) => {
+    const { userId } = req.params
+
+    userService.getById(userId)
+        .then(user => res.send(user))
+        .catch(err => {
+            loggerService.error('Cannot load user', err)
+            res.status(400).send('Cannot load user')
+        })
+})
+
+// Auth API
+app.post('/api/auth/login', (req, res) => {
+    const credentials = req.body
+
+    authService.checkLogin(credentials)
+        .then(user => {
+            const loginToken = authService.getLoginToken(user)
+            res.cookie('loginToken', loginToken)
+            res.send(user)
+        })
+        .catch(() => res.status(404).send('Invalid Credentials'))
+})
+
+app.post('/api/auth/signup', (req, res) => {
+    const credentials = req.body
+
+    userService.add(credentials)
+        .then(user => {
+            if (user) {
+                const loginToken = authService.getLoginToken(user)
+                res.cookie('loginToken', loginToken)
+                res.send(user)
+            } else {
+                res.status(400).send('Cannot signup')
+            }
+        })
+        .catch(err => res.status(400).send('Username taken.'))
+})
+
+app.post('/api/auth/logout', (req, res) => {
+    res.clearCookie('loginToken')
+    res.send('logged-out!')
+})
+
+
+app.get('/**', (req, res) => {
+    res.sendFile(path.resolve('public/index.html'))
+})
+
+app.listen(3030, () => console.log('Server ready at port 3030'))
